requirements.txt
discord.py==0.16.12
asyncio
youtube_dl
PyNaCl